app/requirements.txt
fastapi==0.115.6
funasr==1.1.14
pydantic==2.10.3
pydub==0.25.1
PyYAML==6.0.2
Requests==2.32.3
sparkai==0.3.0
modelscope
websocket-client
onnx
torch
torchaudio
pydub
uvicorn
onnxconverter_common
modelscope
huggingface
huggingface_hub
funasr>=1.1.2
numpy<=1.26.4
gradio
websocket
tenacity